% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrbp_2_rsid.R
\name{chrbp_2_rsid}
\alias{chrbp_2_rsid}
\title{A Function that uses biomaRt to extract the rsid for provided chromosome and base pair.}
\usage{
chrbp_2_rsid(chrbp)
}
\arguments{
\item{chrbp}{chromosome:basepair character}
}
\description{
This function extracts rsids for provided mapping coordinates. Specific for build hg19.
}
\examples{
chrbp_2_rsid()
}
\keyword{annotation}
\keyword{biomart,}
\keyword{map,}
\keyword{rsid,}
