% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetermineK.R
\name{DetermineK}
\alias{DetermineK}
\title{A kmeans and Silhouette Analysis to determine the best K for your data}
\usage{
DetermineK(
  mydata,
  kmax = 20,
  desiredK = 4,
  num_iter = 100,
  plotrepress = FALSE
)
}
\arguments{
\item{mydata}{a data frame or matrix of data}

\item{kmax}{the maximum number of k groups to estimate}

\item{desiredK}{the the number of k clusters you want to have}

\item{num_iter}{the number of iterations used in the kmeans analysis}

\item{plotrepress}{default is to plot the data (FALSE), set to TRUE to repress the plot}
}
\description{
This function estimates the best K (number of clusters) for your data set using Kmeans and a Silhouette Analysis
}
\examples{
DetermineK()
}
\keyword{kmeans,}
\keyword{silhouette}
