% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.sumstats.R
\name{df.sumstats}
\alias{df.sumstats}
\title{A Function to describe each feature in a data frame of metabolite data.}
\usage{
df.sumstats(wdata)
}
\arguments{
\item{wdata}{the metabolite data drame samples in row, metabolites in columns}
}
\description{
This function allows you to generate summary statistics for metabolite features using the describe() function from the psych package, as well as estimate variance, a dispersion index, the coeficent of variation, proportion of zeros in data, and shapiro's W-statistic.
}
\examples{
df.sumstats()
}
\keyword{metabolome,}
\keyword{microbiome}
