% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.test.R
\name{enrich.test}
\alias{enrich.test}
\title{A simple hypergeomeric enrichment test}
\usage{
enrich.test(background, sig)
}
\arguments{
\item{background}{a vector of variables for all tested observations}

\item{sig}{a vector of variables for all "significant" observations}
}
\description{
This function performs a hypergeometric test on each unique class of variable in a list.
}
\examples{
enrich.test()
}
\keyword{exact}
\keyword{fisher}
\keyword{geometric}
\keyword{hyper}
\keyword{test}
\keyword{test,}
