% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_on_factor.R
\name{factor_on_factor}
\alias{factor_on_factor}
\title{A Function to describe each feature in a data frame of metabolite data.}
\usage{
factor_on_factor(x, y)
}
\arguments{
\item{x}{a vector of class factor}

\item{y}{a second vector of class factor}
}
\description{
This function estimates the correlation structure among two factors, using a chi-square test and the Crammer's V statsitic
}
\examples{
factor_on_factor()
}
\keyword{among}
\keyword{analysis}
\keyword{correlation}
\keyword{factors}
