% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_on_numeric.R
\name{factor_on_numeric}
\alias{factor_on_numeric}
\title{A Function to describe each feature in a data frame of metabolite data.}
\usage{
factor_on_numeric(cat_values, num_values)
}
\arguments{
\item{cat_values}{a vector of class factor, categorical values}

\item{num_values}{a second vector of class numeric}
}
\description{
This function estimates the correlation structure among a factor and a numeric|quantitative trait
}
\examples{
factor_on_numeric()
}
\keyword{among}
\keyword{analysis}
\keyword{correlation}
\keyword{factors}
