% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgtest.R
\name{hgtest}
\alias{hgtest}
\title{A Function to to perform a hypergeometric test and a fisher exact test}
\usage{
hgtest(allfeatures, selectedfeatures)
}
\arguments{
\item{allfeatures}{a vector of class categorical, for all elements tested, identifying a category | or bin | box that each element belong to}

\item{selectedfeatures}{a vector of class categorical, for all elements that are unique for some reason, identifying a category | or bin | box that each element belong to}
}
\description{
This function estimates the correlation structure among two factors, using a chi-square test and the Crammer's V statsitic
}
\examples{
hgtest()
}
\keyword{exact}
\keyword{fisher}
\keyword{hypergeometric}
\keyword{test}
\keyword{test;}
