% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moose_barplot.R
\name{moose_barplot}
\alias{moose_barplot}
\title{A Function to make a barplot}
\usage{
moose_barplot(
  data,
  column_to_plot,
  labels_vec,
  rot_angle,
  pcol = "steelblue",
  pmain = "",
  pylim = c(0, 1)
)
}
\arguments{
\item{data}{the data frame holding the data to plot}

\item{column_to_plot}{is the column name to plot}

\item{labels_vec}{is a vector of label names to identify each bar}

\item{rot_angle}{the angle to rotate the x-axis labels}

\item{pcol}{is the plottin color}

\item{pmain}{is the plotting main title}

\item{pylim}{is the plotting y limits}
}
\description{
This function creates a barplot with  rotated x labels
}
\examples{
moose_barplot()
}
\keyword{among}
\keyword{analysis}
\keyword{correlation}
\keyword{factors}
