% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moose_biplot.R
\name{moose_biplot}
\alias{moose_biplot}
\title{A Function to describe each feature in a data frame of metabolite data.}
\usage{
moose_biplot(
  PCA,
  dataframe_of_phenotypes,
  plot_top_N_phenotypes = 3,
  groupfactor
)
}
\arguments{
\item{PCA}{a prcomp object}

\item{dataframe_of_phenotypes}{a dataframe of phenotypes to correlate against PC1 and PC2 of the prcomp object. The function currently only uses quantitative traits in the analysis. Categorical is not possible, yet.}

\item{plot_top_N_phenotypes}{an integer specifying how many of the top correlated phenotypes should be plotted}

\item{groupfactor}{a vector of class categorical, of the sample length, as the number fo data points in PC1 and PC2. This categorical data will aid in color coding your plotted points.}
}
\description{
This function estimates the correlation structure among two factors, using a chi-square test and the Crammer's V statsitic
}
\examples{
moose_biplot()
}
\keyword{biplot}
