% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normW.R
\name{normW}
\alias{normW}
\title{A function to provide a Shapiro-Wilk w-statistic}
\usage{
normW(y)
}
\arguments{
\item{y}{a vector of numerical values}
}
\value{
a vector of lenght 1 providing the W-statistic
}
\description{
This function performs a Shapiro-Wilk test of normality and returns just the W-statistic
}
\examples{
normW()
}
\keyword{normality}
\keyword{shapiro,}
