% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_on_numeric.R
\name{numeric_on_numeric}
\alias{numeric_on_numeric}
\title{A Function to describe each feature in a data frame of metabolite data.}
\usage{
numeric_on_numeric(x, y)
}
\arguments{
\item{x}{a vector of class numeric}

\item{c}{a vector of class numeric}
}
\description{
This function estimates the correlation structure among two factors, using a chi-square test and the Crammer's V statsitic
}
\examples{
numeric_on_numeric()
}
\keyword{among}
\keyword{analysis}
\keyword{correlation}
\keyword{factors}
