% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmr.R
\name{osmr}
\alias{osmr}
\title{A function to perform a one sample MR}
\usage{
osmr(
  wdata,
  outcome,
  exposure,
  instrument,
  covariates = NA,
  outcome_model_family = "gaussian",
  exposure_model_family = "gaussian",
  weights = NA,
  rnt_outcome = FALSE
)
}
\arguments{
\item{wdata}{a data frame of data with appropriate column names}

\item{outcome}{a string defining the outcome (column) name}

\item{exposure}{a string defining the exposure (column) name}

\item{instrument}{a string defining the instrument (column) name}

\item{covariates}{a vector of covariate (column) names}

\item{outcome_model_family}{gaussian or binomial- should the outcome be modeled as a gaussian or binomial}

\item{exposure_model_family}{gaussian or binomial - should the exposure be modeled as a gaussian or binomial}

\item{weights}{if needed, a string defining the weights (column) name}

\item{rnt_outcome}{TRUE or FALSE should the outcome trait be rank normal transformed}
}
\value{
a vector of summary statistics
}
\description{
This function performs a one sample MR, including the observational (oe: outcome on exposure) analysis, the exposure on instrument (ei) analysis, and the Mendelian Randomization (MR) or instrumental variable analysis.
}
\examples{
osmr()
}
\keyword{Hausman}
\keyword{MR,}
\keyword{Wu}
\keyword{one-sample}
\keyword{test}
