% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileNormalization.R
\name{quantile_normalisation}
\alias{quantile_normalisation}
\title{A function to perform a quantile normalisation across a data.frame}
\usage{
quantile_normalisation(df)
}
\arguments{
\item{df}{a data frame of integer or numerical values}
}
\description{
This function performs a standard quantile normalization often seen in things like gene expression data.
}
\examples{
quantile_normalisation()
}
\keyword{normalization}
\keyword{qqnorm,}
\keyword{quantile}
