% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rntransform.R
\name{rntransform}
\alias{rntransform}
\title{A Function to rank normal transform a distribution}
\usage{
rntransform(formula, data, family = gaussian, split_ties = TRUE)
}
\arguments{
\item{formula}{ztransformation formula}

\item{data}{the vector of data to be transformed}

\item{family}{the distribution family used in the ztransformation}

\item{split_ties}{a flag asking if tied values should be split at random or left alone}
}
\description{
This function to rank normal transform a vect or of data
}
\examples{
rntransform()
}
\keyword{among}
\keyword{analysis}
\keyword{correlation}
\keyword{factors}
