% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_2_chrbp.R
\name{rsid_2_chrbp}
\alias{rsid_2_chrbp}
\title{A Function that uses biomaRt to extract chromosome and base pair for your rsid}
\usage{
rsid_2_chrbp(rsid)
}
\arguments{
\item{rsid}{dbSNP reference id}
}
\description{
This function extracts mapping coordinates for your rsid
}
\examples{
rsid_2_chrbp()
}
\keyword{annotation}
\keyword{biomart,}
\keyword{map,}
\keyword{rsid,}
