% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpid_2_geneanno.R
\name{snpid_2_geneanno}
\alias{snpid_2_geneanno}
\title{A Function that uses biomaRt to annotate SNPIDs}
\usage{
snpid_2_geneanno(snpid, plusminus_dist = 150)
}
\arguments{
\item{snpid}{chr:basepair_A1_A2}

\item{plusminus_dist}{plus minus distsance in kilobases (kb) to annotate nearby genes.}
}
\description{
This function extracts mapping and (+/- 150KB ) gene annotation information for SNPids
}
\examples{
snpid_2_geneanno()
}
\keyword{annotation}
\keyword{biomart,}
\keyword{map,}
\keyword{rsid,}
