% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_DF_Correlations.R
\name{Test_DF_Correlations}
\alias{Test_DF_Correlations}
\title{A Function estimate the correlation among all columns in a data frame, categorical and quantitative}
\usage{
Test_DF_Correlations(df)
}
\arguments{
\item{df}{a data frame of data, can contain both numeric (quantitative) and factor (categorical) data}
}
\description{
This function estimates the correlation structure among all columns in a data frame
}
\examples{
Test_DF_Correlations()
}
\keyword{among}
\keyword{analysis}
\keyword{correlation}
\keyword{factors}
