% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.cut.height.mean.rho.R
\name{tree.cut.height.mean.rho}
\alias{tree.cut.height.mean.rho}
\title{A Function to iterate over a variety of tree cut heights to estimate the number of clusters, the min, the mean, the max correlation estimate for all elements with a cluster.}
\usage{
tree.cut.height.mean.rho(hclust_data, CMat)
}
\arguments{
\item{hclust_data}{an hclust object}

\item{CMat}{a correlation matrix used to generate the hclut object}
}
\description{
This function estimates the correlation structure among two factors, using a chi-square test and the Crammer's V statsitic
}
\examples{
tree.cut.height.mean.rho()
}
\keyword{clusters,}
\keyword{correlation}
\keyword{data}
\keyword{estimates}
\keyword{independence}
\keyword{reduction,}
\keyword{variable}
\keyword{within}
