% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplot_col.R
\name{vioplot_col}
\alias{vioplot_col}
\title{A Function edited by Zhisong He to add color to a violin plot}
\usage{
vioplot_col(
  x,
  ...,
  range = 1.5,
  h = NULL,
  ylim = NULL,
  names = NULL,
  horizontal = FALSE,
  col = rep("magenta", length(list(x, ...))),
  border = "black",
  lty = 1,
  lwd = 1,
  rectCol = "black",
  colMed = "white",
  pchMed = 19,
  at,
  add = FALSE,
  wex = 1,
  drawRect = TRUE
)
}
\arguments{
\item{x}{a vector of numerical values}
}
\description{
an edit to vioplot() to add colors to the violin|box plots.
}
\examples{
vioplot_col()
}
\keyword{plot}
\keyword{violin}
