% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztransform.R
\name{ztransform}
\alias{ztransform}
\title{A Function to Z transform a vector of data}
\usage{
ztransform(formula, data, family = gaussian)
}
\arguments{
\item{formula}{ztransformation formula}

\item{data}{the vector of data to be transformed}

\item{family}{the distribution family used in the ztransformation}
}
\description{
A Function to Z transform a vector of data
}
\examples{
ztransform()
}
\keyword{among}
\keyword{analysis}
\keyword{correlation}
\keyword{factors}
